Moving color validate functions to util for use in multiple files
