+++ src/plugins/codeStyle/index.ts
@@ -1,7 +1,7 @@
-import { BscFile, BsDiagnostic, createVisitor, FunctionExpression, isBrsFile, isGroupingExpression, TokenKind, WalkMode, CancellationTokenSource, DiagnosticSeverity, OnGetCodeActionsEvent, isCommentStatement, AALiteralExpression, LiteralExpression, AAMemberExpression } from 'brighterscript';
-import { RuleAAComma, RuleColorFormat, RuleColorCase, RuleColorAlpha, RuleColorAlphaDefaults, RuleColorCertCompliant } from '../..';
+import { BscFile, BsDiagnostic, createVisitor, FunctionExpression, isBrsFile, isGroupingExpression, TokenKind, WalkMode, CancellationTokenSource, DiagnosticSeverity, OnGetCodeActionsEvent, isCommentStatement, AALiteralExpression, AAMemberExpression, isXmlFile } from 'brighterscript';
+import { RuleAAComma } from '../..';
 import { addFixesToEvent } from '../../textEdit';
-import { PluginContext } from '../../util';
+import { PluginContext, validateColorStyle } from '../../util';
 import { messages } from './diagnosticMessages';
 import { extractFixes } from './styleFixes';
 
@@ -18,7 +18,7 @@ export default class CodeStyle {
     }
 
     afterFileValidate(file: BscFile) {
-        if (!isBrsFile(file) || this.lintContext.ignores(file)) {
+        if ((!isBrsFile(file) && !isXmlFile(file)) || this.lintContext.ignores(file)) {
             return;
         }
 
@@ -128,10 +128,8 @@ export default class CodeStyle {
                 }
             },
             LiteralExpression: e => {
-                if (e.token.kind === TokenKind.StringLiteral) {
-                    if (validateColorFormat) {
-                        this.validateColorStyle(e.token, diagnostics, colorFormat, colorCase, colorAlpha, colorAlphaDefaults, colorCertCompliant);
-                    }
+                if (validateColorFormat && e.token.kind === TokenKind.StringLiteral) {
+                    validateColorStyle(e.token, diagnostics, colorFormat, colorCase, colorAlpha, colorAlphaDefaults, colorCertCompliant);
                 }
             },
             AALiteralExpression: e => {
@@ -171,107 +169,6 @@ export default class CodeStyle {
 
         // append diagnostics
         file.addDiagnostics(bsDiagnostics);
-    }
-
-    validateColorStyle(token: LiteralExpression, diagnostics: (Omit<BsDiagnostic, 'file'>)[], colorFormat: RuleColorFormat, colorCase: RuleColorCase, alpha: RuleColorAlpha, alphaDefaults: RuleColorAlphaDefaults, certCompliant: RuleColorCertCompliant) {
-        const colorHashRegex = /#[0-9A-Fa-f]{6}/g;
-        const colorHashAlphaRegex = /#[0-9A-Fa-f]{8}/g;
-        const colorZeroXRegex = /0x[0-9A-Fa-f]{6}/g;
-        const colorZeroXAlphaRegex = /0x[0-9A-Fa-f]{8}/g;
-        const colorHashMatches = token.text.match(colorHashRegex);
-        const colorHashAlphaMatches = token.text.match(colorHashAlphaRegex);
-        const colorZeroXMatches = token.text.match(colorZeroXRegex);
-        const colorZeroXAlphaMatches = token.text.match(colorZeroXAlphaRegex);
-
-        if (colorFormat === 'hash') {
-            if (colorZeroXMatches !== null) {
-                diagnostics.push(messages.expectedColorFormat(token.range));
-            }
-            this.validateColorCase(colorHashMatches, token, diagnostics, colorCase, colorFormat);
-            this.validateColorAlpha(colorHashAlphaMatches, colorHashMatches, colorZeroXMatches, token, diagnostics, alpha, alphaDefaults);
-            this.validateColorCertCompliance(colorHashMatches, token, diagnostics, colorFormat, certCompliant);
-
-        } else if (colorFormat === 'zero-x') {
-            if (colorHashMatches !== null) {
-                diagnostics.push(messages.expectedColorFormat(token.range));
-            }
-            this.validateColorCase(colorZeroXMatches, token, diagnostics, colorCase, colorFormat);
-            this.validateColorAlpha(colorZeroXAlphaMatches, colorHashMatches, colorZeroXMatches, token, diagnostics, alpha, alphaDefaults);
-            this.validateColorCertCompliance(colorZeroXMatches, token, diagnostics, colorFormat, certCompliant);
-
-        } else if (colorFormat === 'never') {
-            if (colorZeroXMatches !== null || colorHashMatches !== null) {
-                diagnostics.push(messages.expectedColorFormat(token.range));
-            }
-        }
-    }
-
-    validateColorAlpha(alphaMatches: RegExpMatchArray, hashMatches: RegExpMatchArray, zeroXMatches: RegExpMatchArray, token: LiteralExpression, diagnostics: (Omit<BsDiagnostic, 'file'>)[], alpha: RuleColorAlpha, alphaDefaults: RuleColorAlphaDefaults) {
-        const validateColorAlpha = (alpha === 'never' || alpha === 'always' || alpha === 'allowed');
-        if (validateColorAlpha) {
-            if (alpha === 'never' && alphaMatches !== null) {
-                diagnostics.push(messages.expectedColorAlpha(token.range));
-            }
-            if ((alpha === 'always' && alphaMatches === null) && (hashMatches !== null || zeroXMatches !== null)) {
-                diagnostics.push(messages.expectedColorAlpha(token.range));
-            }
-            if ((alphaDefaults === 'never' || alphaDefaults === 'only-hidden') && alphaMatches !== null) {
-                for (let i = 0; i < alphaMatches.length; i++) {
-                    const colorHashAlpha = alphaMatches[i];
-                    const alphaValue = colorHashAlpha.slice(-2).toLowerCase();
-                    if (alphaValue === 'ff' || (alphaDefaults === 'never' && alphaValue === '00')) {
-                        diagnostics.push(messages.expectedColorAlphaDefaults(token.range));
-                    }
-                }
-            }
-        }
-    }
-
-    validateColorCase(matches: RegExpMatchArray, token: LiteralExpression, diagnostics: (Omit<BsDiagnostic, 'file'>)[], colorCase: RuleColorCase, colorFormat: RuleColorFormat) {
-        const validateColorCase = colorCase === 'upper' || colorCase === 'lower';
-        if (validateColorCase && matches !== null) {
-            let colorValue = matches[0];
-            const charsToStrip = (colorFormat === 'hash') ? 1 : 2;
-            colorValue = colorValue.substring(charsToStrip);
-            for (let i = 0; i < colorValue.length; i++) {
-                const char = colorValue.charAt(i);
-                if (colorCase === 'lower' && char === char.toUpperCase() && char !== char.toLowerCase()) {
-                    diagnostics.push(messages.expectedColorCase(token.range));
-                    break;
-                }
-                if (colorCase === 'upper' && char === char.toLowerCase() && char !== char.toUpperCase()) {
-                    diagnostics.push(messages.expectedColorCase(token.range));
-                    break;
-                }
-            }
-        }
-    }
-
-    validateColorCertCompliance(matches: RegExpMatchArray, token: LiteralExpression, diagnostics: (Omit<BsDiagnostic, 'file'>)[], colorFormat: RuleColorFormat, certCompliant: RuleColorCertCompliant) {
-        const validateCertCompliant = certCompliant === 'always';
-        if (validateCertCompliant && matches !== null) {
-            const BROADCAST_SAFE_BLACK = '161616';
-            const BROADCAST_SAFE_WHITE = 'DBDBDB';
-            const MAX_BLACK_LUMA = this.getColorLuma(BROADCAST_SAFE_BLACK);
-            const MAX_WHITE_LUMA = this.getColorLuma(BROADCAST_SAFE_WHITE);
-            let colorValue = matches[0];
-            const charsToStrip = (colorFormat === 'hash') ? 1 : 2;
-            colorValue = colorValue.substring(charsToStrip);
-            const colorLuma = this.getColorLuma(colorValue);
-            if (colorLuma > MAX_WHITE_LUMA || colorLuma < MAX_BLACK_LUMA) {
-                diagnostics.push(messages.colorCertCompliance(token.range));
-            }
-        }
-    }
-
-    getColorLuma(value: string) {
-        let luma = -1;
-        const rgb = parseInt(value, 16); // Convert rrggbb to decimal
-        const red = (rgb >> 16) & 0xff;
-        const green = (rgb >> 8) & 0xff;
-        const blue = (rgb >> 0) & 0xff;
-        luma = 0.2126 * red + 0.7152 * green + 0.0722 * blue; // Per ITU-R BT.709
-        return luma;
     }
 
     validateAAStyle(aa: AALiteralExpression, aaCommaStyle: RuleAAComma, diagnostics: (Omit<BsDiagnostic, 'file'>)[]) {
