+++ test/project1/components/colors/color-broadcast-safe.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<component name="ColorBroadcastSafe" extends="Group">
+    <children>
+        <Rectangle color="#DBDBDC"/>
+        <Label color="#161616"/>
+        <SimpleLabel color="#DBDBDBFF"/>
+        <Rectangle color="#161615"/>
+    </children>
+</component>
